+++ src/test/java/seedu/multitasky/commons/util/match/PowerMatchTest.java
@@ -1,4 +1,4 @@
-package seedu.multitasky.commons.util;
+package seedu.multitasky.commons.util.match;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
