+++ src/test/java/seedu/multitasky/logic/commands/RestoreCommandTest.java
@@ -12,7 +12,7 @@ import org.junit.Test;
 
 import seedu.multitasky.commons.core.Messages;
 import seedu.multitasky.commons.core.index.Index;
-import seedu.multitasky.commons.util.PowerMatch;
+import seedu.multitasky.commons.util.match.PowerMatch;
 import seedu.multitasky.logic.CommandHistory;
 import seedu.multitasky.logic.parser.Prefix;
 import seedu.multitasky.model.Model;
